+++ lib/glimmer/dsl/swt/dsl.rb
@@ -49,7 +49,6 @@ module Glimmer
           data_binding
           cursor
           font
-          image
           multiply
           property
           block_property